refactor: move search types and preview component to web-pkg
