+++ packages/web-app-files/src/helpers/resource/index.ts
@@ -1,2 +1,1 @@
-export * from './filter'
 export * from './actions'
