+++ packages/web-app-skeleton/src/search/github/index.ts
@@ -1,6 +1,5 @@
 import Preview from './preview'
-import { SearchPreview, SearchProvider } from 'web-app-search/src/types'
-import { EventBus } from '@ownclouders/web-pkg'
+import { EventBus, SearchPreview, SearchProvider } from '@ownclouders/web-pkg'
 import { GitHubClient } from './client'
 import { HttpClient } from '@ownclouders/web-pkg'
