+++ packages/web-pkg/src/components/Search/index.ts
@@ -1,0 +1,2 @@
+export { default as ResourcePreview } from './ResourcePreview.vue'
+export * from './types'
