+++ packages/web-pkg/src/helpers/resource/filter.ts
@@ -1,5 +1,5 @@
 import Fuse from 'fuse.js'
-import { defaultFuseOptions } from '@ownclouders/web-pkg'
+import { defaultFuseOptions } from '../fuse'
 
 export const filterResources = (resources: unknown[], term: string, limit?: number): unknown[] => {
   const engine = new Fuse(resources, {
